changed internal item types from uint to byte to match N64 type

+++ OcarinaMultiworld.Lib/Items/Item.cs
@@ -3,10 +3,10 @@
     public record Item
     {
         public string Name        { get; }
-        public uint?  SendId      { get; }
-        public uint?  InventoryId { get; }
+        public byte?  SendId      { get; }
+        public byte?  InventoryId { get; }
 
-        internal Item(string name, uint? sendId, uint? inventoryId)
+        internal Item(string name, byte? sendId, byte? inventoryId)
         {
             Name = name;
             SendId = sendId;
